refactor PlayableItem
